FIX ensure TOTP icon does not sit over the header
When verifying a TOTP code, the icon to the right of the input field was shifted
up and sat over the title. This is fine on the front-end, but in the CMS the title
is within a header on a modal, and this looked strange. It was also out of line
(haha, pun intended) with other authenticators and screens on that user flow where
all icons sat below the heading.

Removing the negative margin on the icon containing div solves this, an also makes
up for the removal of the `form-group` class that was adding an undesired divider
line between the input field and the actions, via silverstripe/admin applied styles
(when in the CMS - via `admin/myprofile`). This change removes spacing between the
input and the next button (et. al.), but the image moving down re-introduces the
spacing by forcing the parent container to be a larger height (that the class
mfa-totp__validate-left does not fill).